Enable index scan on AO tables for indexes that don’t have amgetbitmap (#16231)
* Enable index scan on AO tables for indexes that don’t have amgetbitmap

Ivfflat and hsnw indexes both don't support 'amgetbitmap', after this pr
they can be used on AO tables.

The main reason for us to enable those 2 indexes on AO tables is that
they both only support ORDER BY queries and the typical query on
embedding data is "ORDER BY embedding LIMIT", thus only limited rows
fetched from the base tables, which means that the cost of random access
on AO tables won't be big.

The other reason is that the IVFFLAT index only fetches part of the
index tuples when doing a query (the percent is probes/lists), so
generally using the index can improve the performance.

No such indexes (amgetbitmap=NULL but amgettuple=non-NULL) exist in core
or have ever been supported for GPDB customers, so no chance of
regression with this change.