Fix flaky test case vacuum_stats
During the test, backgroup process like dtx recover and FTS will
create gang, which will start transaction on segment. And concurrent
transacion will affect the behavior of vacuum (OldestXmin), which will
result in that relallvisible could be 0.

In this commit, we make sure that dtx recover and FTS are not working
in the test.